Move the base controls to a "BaseUserControls" namespace
