+++ Bardez.Project.SwordOfTheStars.Editor.UI.WinForms/UserControls/BaseUserControls/UserControlBase.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows.Forms;
 using UC = System.Windows.Forms.UserControl;
 
-namespace Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls
+namespace Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.BaseUserControls
 {
     public abstract class UserControlBase : UC
     {
@@ -12,9 +12,9 @@ namespace Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls
 
         protected void SetStyles()
         {
-            this.SetStyle(ControlStyles.OptimizedDoubleBuffer | ControlStyles.AllPaintingInWmPaint | ControlStyles.UserPaint, true);
-            this.SetStyle(ControlStyles.EnableNotifyMessage, true);
-            this.DoubleBuffered = true;
+            SetStyle(ControlStyles.OptimizedDoubleBuffer | ControlStyles.AllPaintingInWmPaint | ControlStyles.UserPaint, true);
+            SetStyle(ControlStyles.EnableNotifyMessage, true);
+            DoubleBuffered = true;
         }
 
         protected override void OnNotifyMessage(Message m)
