+++ Bardez.Project.SwordOfTheStars.Editor.UI.WinForms/UserControls/Players/Players.cs
@@ -4,7 +4,7 @@ using System.Data;
 using System.Windows.Forms;
 using Bardez.Project.SwordOfTheStars.DataStructures;
 using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.Helpers;
-using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.UserControl;
+using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.BaseUserControls;
 using Bardez.Project.SwordOfTheStars.UI.Abstractions.TechTree.Graph;
 
 namespace Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.Players
