+++ Bardez.Project.SwordOfTheStars.Editor.UI.WinForms/UserControls/Species/SpeciesDetails.cs
@@ -2,7 +2,7 @@ using System;
 using System.Data;
 using System.Windows.Forms;
 using Bardez.Project.SwordOfTheStars.DataStructures;
-using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.UserControl;
+using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.BaseUserControls;
 
 namespace Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.Species
 {
