+++ Bardez.Project.SwordOfTheStars.Editor.UI.WinForms/UserControls/Systems/SystemDetails.cs
@@ -1,6 +1,6 @@
 using System;
 using Bardez.Project.SwordOfTheStars.DataStructures;
-using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.UserControl;
+using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.BaseUserControls;
 
 namespace Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls
 {
