+++ Bardez.Project.SwordOfTheStars.Editor.UI.WinForms/UserControls/Systems/SystemDetailsHistory.cs
@@ -1,5 +1,5 @@
 using Bardez.Project.SwordOfTheStars.DataStructures;
-using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.UserControl;
+using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.BaseUserControls;
 
 namespace Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls
 {
