+++ Bardez.Project.SwordOfTheStars.Editor.UI.WinForms/UserControls/Systems/Systems.cs
@@ -3,7 +3,7 @@ using System.Data;
 using System.Windows.Forms;
 using Bardez.Project.SwordOfTheStars.DataStructures;
 using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.Helpers;
-using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.UserControl;
+using Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls.BaseUserControls;
 
 namespace Bardez.Project.SwordOfTheStars.Editor.UI.WinForms.UserControls
 {
